requirements.txt
scikit-learn==1.1.1
pandas==1.4.2
pytorch==1.11.0
numpy==1.22.3
